pyproject.toml
[project]
name = "notebookllama"
version = "0.4.0"
description = "An OSS and LlamaCloud-backed alternative to NotebookLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "audioop-lts>=0.2.1",
  "cohere>=5.15.0",
  "elevenlabs>=2.5.0",
  "fastmcp>=2.9.2",
  "ffprobe>=0.5",
  "llama-cloud>=0.1.29",
  "llama-cloud-services>=0.6.38",
  "llama-index-core>=0.12.44",
  "llama-index-embeddings-azure-inference>=0.3.0",
  "llama-index-embeddings-bedrock>=0.5.2",
  "llama-index-embeddings-cohere>=0.5.1",
  "llama-index-embeddings-gemini>=0.3.2",
  "llama-index-embeddings-huggingface-api>=0.3.1",
  "llama-index-embeddings-openai>=0.3.1",
  "llama-index-indices-managed-llama-cloud>=0.6.11",
  "llama-index-llms-cohere>=0.5.0",
  "llama-index-llms-openai>=0.4.7",
  "llama-index-observability-otel>=0.1.1",
  "llama-index-tools-mcp>=0.2.5",
  "llama-index-workflows>=1.0.1",
  "markdown-analysis>=0.1.5",
  "mypy>=1.16.1",
  "opentelemetry-exporter-otlp-proto-http>=1.34.1",
  "plotly>=6.2.0",
  "pre-commit>=4.2.0",
  "psycopg2-binary>=2.9.10",
  "pydub>=0.25.1",
  "pytest>=8.4.1",
  "pytest-asyncio>=1.0.0",
  "python-dotenv>=1.1.1",
  "pyvis>=0.3.2",
  "randomname>=0.2.1",
  "streamlit>=1.46.1",
  "textual>=3.7.1",
]

[tool.mypy]
disable_error_code = ["import-not-found", "import-untyped"]

[tool.pytest.ini_options]
pythonpath = ["src"]
